refactor: move transaction and generate page
