+++ ui/modals/generate/cmds.go
