+++ ui/modals/generate/controller.go
