+++ ui/modals/generate/model.go
