+++ ui/modals/generate/view.go
