+++ ui/modals/transaction/model.go
