+++ ui/modals/transaction/view.go
